Move fonts to separate go file
