+++ main.go
@@ -2,14 +2,10 @@ package main
 
 import (
 	"embed"
-	_ "image/png"
 	"log"
 
 	"github.com/hajimehoshi/ebiten/v2"
-	"github.com/hajimehoshi/ebiten/v2/examples/resources/fonts"
 	"github.com/hajimehoshi/ebiten/v2/inpututil"
-	"golang.org/x/image/font"
-	"golang.org/x/image/font/opentype"
 )
 
 const (
@@ -17,50 +13,8 @@ const (
 	windowHeight = 960
 )
 
-var (
-	mplusBigFont    font.Face
-	mplusNormalFont font.Face
-	mplusSmallFont  font.Face
-	//go:embed assets/catch/*.json assets/*.png
-	assets embed.FS
-)
-
-func init() {
-	tt, err := opentype.Parse(fonts.MPlus1pRegular_ttf)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	const dpi = 72
-	mplusBigFont, err = opentype.NewFace(tt, &opentype.FaceOptions{
-		Size:    36,
-		DPI:     dpi,
-		Hinting: font.HintingFull,
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	mplusNormalFont, err = opentype.NewFace(tt, &opentype.FaceOptions{
-		Size:    24,
-		DPI:     dpi,
-		Hinting: font.HintingFull,
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	mplusSmallFont, err = opentype.NewFace(tt, &opentype.FaceOptions{
-		Size:    20,
-		DPI:     dpi,
-		Hinting: font.HintingFull,
-	})
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-}
+//go:embed assets/catch/*.json assets/*.png
+var assets embed.FS
 
 func setupEnvironment() {
 	ebiten.SetWindowTitle("Magnet Fishing")
